[MEDI-67] fix : edit dto of report
