--- src/main/java/com/my_medi/api/report/dto/BloodPressureDto.java
@@ -1,17 +1,0 @@
-package com.my_medi.api.report.dto;
-
-import com.my_medi.domain.report.enums.BloodPressureStatus;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class BloodPressureDto {
-    private Integer systolic;
-    private Integer diastolic;
-    private BloodPressureStatus bloodPressureStatus;
-}
