+++ src/main/java/com/my_medi/api/report/dto/EditReportRequestDto.java
@@ -1,5 +1,6 @@
 package com.my_medi.api.report.dto;
 
+import com.my_medi.api.report.dto.ReportPartitionRequestDto.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
